Fix scom address mangling for POWER9 type 1 indirect accesses
Skiboot recently added support for type 1 indirect scom addresses.
skiboot commit is here:

  commit 74311a97190d33e3753dc09a5fe6064120c8363a
  Author: Michael Neuling <mikey@neuling.org>
  xscom: Add indirect form 1 scoms

This now uses the whole top nibble of the scom address to determine
the indirection type.

As a consequence, the bit mangling for the kernel interface needs to
be updated (although it's backwards compatible) due to kernel commit.

  commit 517c27570cf38f182e7a688d50a9b978333f8ea8
  Author: Michael Neuling <mikey@neuling.org>
  powerpc/powernv: Fix XSCOM address mangling for form 1 indirect

This borrows the code from skiboot scom utils to add address
mangling. Code is from skiboot commit:

  commit 034d545b1503bf3980c290c361ed1a99ce9d308d
  Author: Michael Neuling <mikey@neuling.org>
  external: Update xscom utils for type 1 indirect accesses

Signed-off-by: Michael Neuling <mikey@neuling.org>